also moved tests to correct package
