+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/handler/applicationservice/processor/readhelper/LazyProxyInputStreamTest.java
@@ -1,4 +1,4 @@
-package com.sap.cds.feature.attachments.handler.applicationservice.processor.applicationevents.model;
+package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
